Updates to code
Updates to code including a player 2 side of things. Changing the control scheme's in the inputs so that two sides of the keyboard can work on the same game. Changing camera angles so that the game now has split screen. And also making it so the car's only move when pressing forward instead of just a constant speed.